+++ app/assets/stylesheets/admin.scss/products.scss
@@ -1,3 +1,5 @@
 // Place all the styles related to the admin::products controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
+@import "bootstrap";
+@import "bootstrap-sprockets";
