+++ app/assets/stylesheets/black.scss
@@ -1,0 +1,13 @@
+.showproduct {
+    background-color:#777777;
+    color:#f8f8f8 !important
+  }
+
+.showproduct-font {
+      color:#f8f8f8 !important
+    }
+
+
+.showproduct-font a {
+      color:#f8f8f8 !important
+    }
