+++ app/controllers/admin/user_mg_controller.rb
@@ -1,6 +1,6 @@
 class Admin::UserMgController < ApplicationController
   before_action :authenticate_user!
-
+  before_action :require_is_admin
 
   def index
     @users=User.all
