+++ app/controllers/welcome_controller.rb
@@ -1,0 +1,5 @@
+class WelcomeController < ApplicationController
+  def index
+
+  end
+end
