+++ app/helpers/welcome_helper.rb
@@ -1,0 +1,2 @@
+module WelcomeHelper
+end
