+++ app/views/admin/products/_form.html.erb
@@ -1,10 +1,13 @@
-<%= simple_form_for [:admin,@product] do |f| %>
-<div class="form-group">
-  <%= f.input :title, input_html:{class:"form-control"} %>
-  <%= f.input :description, input_html:{class:"form-control"} %>
-  <%= f.input :quantity, input_html:{class:"form-control"} %>
-  <%= f.input :price, input_html:{class:"form-control"} %>
-  <%= f.input :image, input_html:{class:"form-control"} %>
+<div style="margin-left:10px">
+  <%= simple_form_for [:admin,@product] do |f| %>
+  <div class="form-group">
+    <%= f.input :title, input_html:{class:"form-control"} %>
+    <%= f.input :description, input_html:{class:"form-control"} %>
+    <%= f.input :quantity, input_html:{class:"form-control"} %>
+    <%= f.input :price, input_html:{class:"form-control"} %>
+    <%= f.input :image, input_html:{class:"form-control"} %>
+  </div>
+    <%= f.submit "submit", class:"btn btn-primary"%>
+  <% end %>
+
 </div>
-  <%= f.submit "submit" %>
-<% end %>
