+++ app/views/admin/products/index.html.erb
@@ -1,18 +1,18 @@
 <div class="col-md-12">
-  <%= link_to("Add a product",new_admin_product_path,class:"btn btn-default") %>
+  <%= link_to("Add a product",new_admin_product_path,class:"btn btn-primary", style:"margin-bottom:10px; background-color:#ea010f") %>
 </div>
 
-<div class="col-md-12">
-<div class="row">
+
+<div class="row" style="margin-left:1px">
 <% @products.each do |product| %>
 
-  <div class="col-sm-5 col-md-2">
-    <div class="thumbnail">
-      <%= image_tag(product.image.medium.url) %>
-      <div class="caption">
+  <div class="col-sm-2 col-md-2">
+    <div class="thumbnail showproduct">
+      <%= link_to(image_tag(product.image.thumb.url),admin_product_path(product)) %>
+      <div class="caption showproduct-font">
         <h5><%= link_to(product.title,admin_product_path(product)) %></h5>
         <p>
-          <%= link_to(product.price,admin_product_path(product)) %>
+          Price: <%= product.price %>
         </p>
         <p><%= link_to("Edit",edit_admin_product_path(product)) %> ｜
           <%= link_to("Delete",admin_product_path(product), method: :delete) %>
@@ -23,5 +23,4 @@
   </div>
 
 <% end %>
-</div>
 </div>
