+++ app/views/admin/products/new.html.erb
@@ -1,2 +1,2 @@
-<h1>Add products</h1>
+<h1 style="margin-left: 10px">Add products</h1>
 <%= render "form" %>
