+++ app/views/common/_navbar.html.erb
@@ -25,6 +25,7 @@
                         <li><%= link_to("管理产品", admin_products_path) %></li>
                         <li><%= link_to("管理用户", admin_user_mg_index_path()) %></li>
                         <% end %>
+                        <li role="separator" class="divider"></li>
                         <li><%= link_to("Logout", destroy_user_session_path, method: :delete) %></li>
                     </ul>
                 </li>
