+++ app/views/products/index.html.erb
@@ -2,10 +2,10 @@
 <div class="row">
 <% @products.each do |product| %>
 
-  <div class="col-sm-5 col-md-3">
-    <div class="thumbnail">
-      <%= image_tag(product.image.thumb.url) %>
-      <div class="caption">
+  <div class="col-sm-2 col-md-2">
+    <div class="thumbnail showproduct">
+      <%= link_to(image_tag(product.image.thumb.url),product_path(product)) %>
+      <div class="caption showproduct-font">
         <h5><%= link_to(product.title,product_path(product)) %></h5>
         <p>
             <%= product.price %>
