+++ app/views/products/show.html.erb
@@ -1,37 +1,32 @@
-<div class="col-md-12">
-  <%= link_to("Back to index",products_path,class:"btn btn-default") %>
-</div>
-<div class="col-md-12">
-  <div class="row">
 
-    <div class="col-sm-5 col-md-8">
-      <div class="thumbnail">
-        <!-- <img src="..." alt="..."> -->
-        <div class="caption">
-          <h5><%= @product.title %></h5>
-          <table>
-            <tr>
-              <td>
-                <%= image_tag @product.image.medium %>
-              </td>
-              <td>
-                <table>
-                    <tr>
-                      <%= @product.description %><br/>
-                      <%= @product.quantity %><br/>
-                      <%= @product.price %><br/>
-                    </tr>
-                </table>
+  <div  class="col-md-12">
+    <%= link_to("Back to index",products_path,class:"btn btn-primary",style:"margin-bottom:10px; background-color:#ea010f") %>
+  </div>
 
-              </td>
-            </tr>
-          </table>
+<div class="col-md-12" >
+  <div class="row" >
+    <div class="col-sm-10 col-md-10" >
+      <!-- <div class="showproduct"> -->
+        <div class="thumbnail showproduct" >
+          <!-- <img src="..." alt="..."> -->
+          <div class="caption showproduct-font" >
+            <div class="media-left">
+              <%= image_tag @product.image.medium %>
+            </div>
 
+            <div class="media-body">
+              <div style=" margin-left: 30px">
+                  <h2 class="media-heading"><%= @product.title %></h2>
+                  Description: <%= @product.description %><br/>
+                  Quantity: <%= @product.quantity %><br/>
+                  Price: <%= @product.price %><br/>
+              </div>
 
-          <!-- <p><a href="#" class="btn btn-primary" role="button">Button</a> <a href="#" class="btn btn-default" role="button">Button</a></p> -->
+            </div>
+          </div>
         </div>
-      </div>
-    </div>
+      <!-- </div> -->
 
+    </div>
   </div>
 </div>
