+++ app/views/welcome/index.html.erb
@@ -1,0 +1,3 @@
+<div style="background-color:#292929" >
+<h1>hello--s</h1>
+</div>
